Everything works, Before about & closure

+++ HtmlDataProcessor.h
@@ -2,4 +2,4 @@ int totalSnapshotDlls(struct processinformation* snapHead);
 unsigned long memoryUsageAvg(struct snapshotsList* head);
 unsigned long totalMemoryUsageCount(struct processinformation* snapHead);
  processinformation* SortByLoadedDlls(struct processinformation* snapHead);
- int highestMemoryUsage(struct processinformation* snapHead);
+ unsigned long long highestMemoryUsage(struct processinformation* snapHead);
